Consolidate dictionary encoding code
