+++ src/parquet/util/CMakeLists.txt
@@ -63,7 +63,6 @@ endif()
 
 ADD_PARQUET_TEST(bit-util-test)
 ADD_PARQUET_TEST(buffer-test)
-ADD_PARQUET_TEST(dict-test)
 ADD_PARQUET_TEST(mem-pool-test)
 ADD_PARQUET_TEST(output-test)
 ADD_PARQUET_TEST(rle-test)
